Updates for improved keyboard handling:
Replaced keyEvent() and unicodeInput() with a new function, keyboardEvent(), that  offers a superset of the functionality of both, and adds the ability to pass in native scan code, virtual key, and modifiers.

Removed the LL_DOM_VK defines, which only ever really made sense for Mozilla.  Replaced them with a KEY_ enum inside the LLQtWebKit class which replicates the keycodes used in the Second Life viewer.

Moved the LinkTargetType constants into an enum inside LLQtWebKit to reduce namespace clutter.

Updated the testgl and ubrowser samples to use the new keyboard input function.