Added a way to handle javascript window.open() reasonably.
Added a new WindowOpenBehavior WOB_SIMULATE_BLANK_HREF_CLICK, which makes actions that would open a new window appear to the host as if they were clicks on links with a target attribute of "_blank" (by sending an appropriate onClickLinkHref message).

Added the LLWebPageOpenShim class, needed to implement the new window open behavior.

Reworked the logic in LLWebPage::acceptNavigationRequest so that all _external and _blank links _only_ send an event and don't do an internal navigate.  This will allow the host to completely define the policy on such links.  This means that the WindowOpenBehavior no longer affects clicks on _external and _blank links, only javascript window.open and clicks with target attributes that would normally open a new window.

Reorganized the Xcode project file list a bit and added some missing header files so that searching the project is more fruitful.

Added a .hgignore file to clean up the output of 'hg status' for my usual workflow.